Rename swing package
