+++ src/main/java/tools/sctrade/companion/CompanionApplication.java
@@ -12,8 +12,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.scheduling.annotation.EnableScheduling;
+import tools.sctrade.companion.gui.CompanionGui;
 import tools.sctrade.companion.input.KeyListener;
-import tools.sctrade.companion.swing.CompanionGui;
 
 @SpringBootApplication
 @EnableScheduling
